+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.0.0 (upcoming)
 
 - Added link field
+- Added group field
 - Removed hide on screen helper
 
 ## 1.0.0 (released 2017-05-30)